refactor: s3 관련 로직을 global 패키지로 이동
